refactor : 학번 폴더 생성
